Unify IntrospectionEndpoint and CivetWrapper
